feat view count

+++ src/main/java/com/sundingyi/forum/dto/QuestionDTO.java
@@ -11,9 +11,9 @@ public class QuestionDTO {
     private long gmtCreate;
     private long gmtModified;
     private Integer creator;
-    private long commentCount;
-    private long viewCount;
-    private long likeCount;
+    private Integer commentCount;
+    private Integer viewCount;
+    private Integer likeCount;
     private String tag;
     private User user;
 }
